more logic to useSearchFilter
